fix types of ammount from double to bigdecimal

+++ back/src/main/java/com/raxrot/back/models/Order.java
@@ -6,6 +6,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
@@ -35,7 +36,7 @@ public class Order {
     @JoinColumn(name = "payment_id")
     private Payment payment;
 
-    private Double totalAmount;
+    private BigDecimal totalAmount;
     private String orderStatus;
 
     @ManyToOne
