+++ back/src/main/java/com/raxrot/back/models/OrderItem.java
@@ -5,6 +5,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Entity
 @Data
 @Table(name = "order_items")
@@ -25,6 +27,6 @@ public class OrderItem {
     private Order order;
 
     private Integer quantity;
-    private double orderedProductPrice;
+    private BigDecimal orderedProductPrice;
 
 }
